added the first advanced task for more functions project

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,8 +7,8 @@
 
 int main(void)
 {
-	long long n = 612852475143;
-	long long i, maxPrime;
+	long n = 612852475143;
+	long i, maxPrime;
 
 	while (n % 2 == 0)
 	{
